chore(dal): remove old emit module
- Remove old emit module since it is unused due to the new pinga-related
  changes
- Consolidate remaining logic into "implicit_emit"
